refactor login component
