+++ frontend/AngularApp/src/app/login/login.component.html
@@ -5,7 +5,26 @@
     </div>
 
     <app-card>
-      <app-login-form></app-login-form>
+      <form class="form" #form="ngForm" (ngSubmit)="onSubmit(form)">
+        <h1 class="title">Login</h1>
+        <div class="form-group">
+          <label for="username">Username</label>
+          <input type="text" id="username" name="username" required ngModel />
+        </div>
+        <div class="form-group">
+          <label for="password">Password</label>
+          <input
+            type="password"
+            id="password"
+            name="password"
+            required
+            ngModel
+          />
+        </div>
+        <button custom-button class="btn">Login</button>
+        <p>Don't have an account yet?</p>
+        <a routerLink="/register">Register</a>
+      </form>
     </app-card>
   </main>
 </div>
