+++ frontend/AngularApp/src/app/register-profile/register-profile.component.html
@@ -1,49 +1,55 @@
 <div class="app-wrapper">
-    <main class="main">
-      <div class="content">
-        <h1 class="title">Sip, Swipe, Spark!</h1>
-      </div>
-  
-      <app-card>
-        <form class="form" #form="ngForm" (ngSubmit)="onSubmit(form)">
-          <h1 class="title">Profile Details</h1>
-          
-          <div class="form-group">
-            <label for="gender">Gender</label>
-            <select id="gender" name="gender" required ngModel>
-              <option value="" disabled selected>Select your gender</option>
-              <option value="male">Male</option>
-              <option value="female">Female</option>
-            </select>
-          </div>
-          
-          <div class="form-group">
-            <label for="preferences">Sexual Preferences</label>
-            <select id="preferences" name="preferences" required ngModel>
-              <option value="" disabled selected>Select your preference</option>
-              <option value="heterosexual">Heterosexual</option>
-              <option value="homosexual">Homosexual</option>
-              <option value="bisexual">Bisexual</option>
-            </select>
-          </div>
-          
-          <div class="form-group">
-            <label for="bio">Biography</label>
-            <textarea id="bio" name="bio" rows="4" required ngModel></textarea>
-          </div>
-          
-          <div class="form-group">
-              <label for="interests">Select Interests</label>
-              <div class="checkbox-group">
-                <div *ngFor="let tag of tagsList">
-                  <input type="checkbox" id="{{ tag }}" name="interests" [value]="tag" (change)="onTagChange($event)" />
-                  <label for="{{ tag }}">{{ tag }}</label>
-                </div>
-              </div>
+  <main class="main">
+    <div class="content">
+      <h1 class="title">Sip, Swipe, Spark!</h1>
+    </div>
+
+    <app-card>
+      <form class="form" #form="ngForm" (ngSubmit)="onSubmit(form)">
+        <h1 class="title">Profile Details</h1>
+
+        <div class="form-group">
+          <label for="gender">Gender</label>
+          <select id="gender" name="gender" required ngModel>
+            <option value="" disabled selected>Select your gender</option>
+            <option value="male">Male</option>
+            <option value="female">Female</option>
+          </select>
+        </div>
+
+        <div class="form-group">
+          <label for="preferences">Sexual Preferences</label>
+          <select id="preferences" name="preferences" required ngModel>
+            <option value="" disabled selected>Select your preference</option>
+            <option value="heterosexual">Heterosexual</option>
+            <option value="homosexual">Homosexual</option>
+            <option value="bisexual">Bisexual</option>
+          </select>
+        </div>
+
+        <div class="form-group">
+          <label for="bio">Biography</label>
+          <textarea id="bio" name="bio" rows="4" required ngModel></textarea>
+        </div>
+
+        <div class="form-group">
+          <label for="interests">Select Interests</label>
+          <div class="checkbox-group">
+            <div *ngFor="let tag of tagsList">
+              <input
+                type="checkbox"
+                id="{{ tag }}"
+                name="interests"
+                [value]="tag"
+                (change)="onTagChange($event)"
+              />
+              <label for="{{ tag }}">{{ tag }}</label>
             </div>
-          
-          <button custom-button class="btn">Save Profile</button>
-        </form>
-      </app-card>
-    </main>
-  </div>
+          </div>
+        </div>
+
+        <button custom-button class="btn">Save Profile</button>
+      </form>
+    </app-card>
+  </main>
+</div>
