+++ frontend/AngularApp/src/app/shared/models/tags.ts
@@ -1,1 +1,22 @@
-export const  tags = ['#travel', '#fitness', '#music', '#photography', '#gaming', '#art', '#yoga', '#reading', '#movies', '#cooking', '#hiking', '#technology', '#fashion', '#nature', '#meditation', '#tattoos', '#cats', '#dogs', '#dance', '#coffee'];
+export const tags = [
+  '#travel',
+  '#fitness',
+  '#music',
+  '#photography',
+  '#gaming',
+  '#art',
+  '#yoga',
+  '#reading',
+  '#movies',
+  '#cooking',
+  '#hiking',
+  '#technology',
+  '#fashion',
+  '#nature',
+  '#meditation',
+  '#tattoos',
+  '#cats',
+  '#dogs',
+  '#dance',
+  '#coffee',
+];
